requirements.windows.txt
cffi==1.14.5
chardet==4.0.0
coloredlogs==15.0
datefinder==0.7.1
dateparser==1.0.0
gitdb==4.0.7
GitPython==3.1.18
hachoir==3.1.2
humanfriendly==9.1
img2pdf==0.4.0
importlib-metadata==4.0.1
lxml==4.6.3
numpy==1.19.5
ocrmypdf==11.7.3
pandas==1.1.5
pdfminer.six==20200517
pdfplumber==0.5.27
pikepdf==2.11.1
Pillow==8.2.0
pluggy==0.13.1
pycparser==2.20
pycryptodome==3.10.1
PyPDF2==1.26.0
pyreadline==2.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
reportlab==3.5.67
six==1.15.0
smmap==4.0.0
sortedcontainers==2.3.0
tqdm==4.60.0
typing-extensions==3.10.0.0
tzlocal==2.1
Wand==0.6.6
XlsxWriter==1.3.9
zipp==3.4.1
